feat(unit-test): add unit tests for reading a file to a slice of bytes
