--- internal/wc_utils/byte_count.go
@@ -1,11 +1,0 @@
-package wcutils
-
-func CountBytes(bytes []byte) int {
-	var byteCount int
-
-	for i := 0; i < len(bytes); i++ {
-		byteCount++
-	}
-
-	return byteCount
-}
