--- internal/wc_utils/read_file.go
@@ -1,15 +1,0 @@
-package wcutils
-
-import (
-	"os"
-)
-
-func FileToBytes(fileName string) []byte {
-	bytes, err := os.ReadFile(fileName)
-	
-	if err != nil {
-		panic(err)
-	}
-
-	return bytes
-}
